Add CollectionReplace Unify

+++ AnyText/Tests/AnyText.Tests/Synchronization/UnifyTests.cs
@@ -3,24 +3,41 @@ using AnyText.Tests.Synchronization.Metamodel.StateMachine;
 using NMF.AnyText;
 using NUnit.Framework;
 
-
 namespace AnyText.Tests.Synchronization
 {
-    
     [TestFixture]
     public class UnifyTests
     {
         private string _tempFilePath;
-        private StateMachineGrammar _grammar;
+        private Parser _parser;
+        private IStateMachine _stateMachine;
+        private SynchronizationService _service;
 
-     
+       
+        [SetUp]
+        public void Setup()
+        {
+            var grammar = new StateMachineGrammar();
+            _tempFilePath = Path.Join(Path.GetTempPath(), Guid.NewGuid().ToString());
+            
+            _parser = grammar.CreateParser();
+            _parser.Context.UsesSynthesizedModel = true;
+
+            _stateMachine = CreateBasicStateMachine();
+            var synthesizedText = grammar.Root.Synthesize(_stateMachine, _parser.Context);
+            File.WriteAllText(_tempFilePath, synthesizedText);
+
+            var synthesizedRootApp = grammar.Root.Synthesize(_stateMachine, new ParsePosition(0, 0), _parser.Context);
+            _parser.UnifyInitialize(synthesizedRootApp, synthesizedText, new Uri(_tempFilePath));
+
+            _service = new SynchronizationService(null, null);
+            _service.SubscribeToModelChanges(_stateMachine, _parser);
+        }
+        
         [TearDown]
         public void TearDown()
         {
-            if (File.Exists(_tempFilePath))
-            {
-                File.Delete(_tempFilePath);
-            }
+            if (File.Exists(_tempFilePath)) File.Delete(_tempFilePath);
         }
 
         private IStateMachine CreateBasicStateMachine()
@@ -30,170 +47,181 @@ namespace AnyText.Tests.Synchronization
             var stateA = new State { Name = "StateA", IsStartState = true };
             var stateB = new State { Name = "StateB" };
             var stateC = new State { Name = "StateC", IsEndState = true };
-
+            var stateD = new State { Name = "StateD" };
             sm.States.Add(stateA);
             sm.States.Add(stateB);
             sm.States.Add(stateC);
+            sm.States.Add(stateD);
 
-            var transition1 = new Transition { Input= "ToB", StartState = stateA, EndState = stateB };
-            var transition2 = new Transition { Input= "ToC", StartState= stateB, EndState = stateC };
+            var transition1 = new Transition { Input = "ToB", StartState = stateA, EndState = stateB };
+            var transition2 = new Transition { Input = "ToC", StartState = stateB, EndState = stateC };
 
             sm.Transitions.Add(transition1);
             sm.Transitions.Add(transition2);
 
             return sm;
         }
-        
-        [SetUp]
-        public void Setup()
-        {
-            _grammar = new StateMachineGrammar();
-            _tempFilePath = Path.Join(Path.GetTempPath(), Guid.NewGuid().ToString());
-        }
-
-        private (IStateMachine RootStateMachine, Parser Parser) SetupStateMachineParser()
-        {
-            var parser = _grammar.CreateParser();
-            parser.Context.UsesSynthesizedModel = true;
-            var sm = CreateBasicStateMachine();
-            var synthesis = _grammar.Root.Synthesize(sm, parser.Context);
-
-            File.WriteAllText(_tempFilePath, synthesis);
-
-
-            var synthesizedRootAppForInit = _grammar.Root.Synthesize(sm, new ParsePosition(0, 0), parser.Context);
-            parser.UnifyInitialize(synthesizedRootAppForInit, synthesis, new Uri(_tempFilePath));
-
-            var service = new SynchronizationService(null, null);
-            service.SubscribeToModelChanges(sm, parser);
-            return (sm, parser);
-        }
 
         [Test]
         public async Task TestDeleteAsync()
         {
-            var (sm, parser) = SetupStateMachineParser();
 
-            var stateToDelete = sm.States.Last(); // Deleting StateC
-            var originalStateCount = sm.States.Count;
-            
+            var stateToDelete = _stateMachine.States.First(s => s.Name == "StateC"); // Deleting StateC
+            var originalStateCount = _stateMachine.States.Count;
+
             stateToDelete.Delete();
-            await Task.Delay(50);
+            await Task.Delay(500);
 
-            IStateMachine parserRootSm = (IStateMachine)parser.Context.Root;
+            var parserRootSm = (IStateMachine)_parser.Context.Root;
             Assert.That(parserRootSm.States.Count, Is.EqualTo(originalStateCount - 1));
             Assert.That(parserRootSm.States.Any(s => s.Name == stateToDelete.Name), Is.False);
 
-            parser.Context.TryGetDefinition(stateToDelete, out var def);
+            _parser.Context.TryGetDefinition(stateToDelete, out var def);
             Assert.That(def, Is.Null);
 
-            Assert.That(parserRootSm, Is.EqualTo(sm));
-
-            Assert.That(parser.Context.Errors.Count(), Is.EqualTo(1)); //Transition loses Reference
-
+            Assert.That(parserRootSm, Is.EqualTo(_stateMachine));
+            Assert.That(_parser.Context.Errors.Count(), Is.EqualTo(1), "One error expected due to a broken transition reference.");
         }
 
         [Test]
         public async Task TestChangeStateNameAsync()
         {
-            var (sm, parser) = SetupStateMachineParser();
 
-            var stateToModify = sm.States.First(); // StateA
+            var stateToModify = _stateMachine.States.First(); // StateA
             var oldName = stateToModify.Name;
             var newName = "ModifiedStateA";
 
             stateToModify.Name = newName;
-            await Task.Delay(50);
+            await Task.Delay(100);
 
-            IStateMachine parserRootSm = (IStateMachine)parser.Context.Root;
+            var parserRootSm = (IStateMachine)_parser.Context.Root;
             var modifiedStateInParser = parserRootSm.States.FirstOrDefault(s => s.Name == newName);
             Assert.That(modifiedStateInParser, Is.Not.Null);
             Assert.That(modifiedStateInParser, Is.SameAs(stateToModify));
             Assert.That(parserRootSm.States.Any(s => s.Name == oldName), Is.False);
 
-            parser.Context.TryGetDefinition(stateToModify, out var def);
+            _parser.Context.TryGetDefinition(stateToModify, out var def);
             Assert.That(def, Is.Not.Null);
             Assert.That(def.GetFirstInnerLiteral().Literal, Is.EqualTo(newName));
             Assert.That(def.ContextElement, Is.SameAs(stateToModify));
 
 
-            Assert.That(parserRootSm, Is.EqualTo(sm));
-            Assert.That(parser.Context.Errors, Is.Empty);
+            Assert.That(parserRootSm, Is.EqualTo(_stateMachine));
+            Assert.That(_parser.Context.Errors, Is.Empty);
         }
 
         [Test]
         public async Task TestChangeTransitionTargetAsync()
         {
-            var (sm, parser) = SetupStateMachineParser();
 
-            var transitionToModify = sm.Transitions.First(t => t.Input == "ToB"); // Transition from StateA to StateB
+            var transitionToModify = _stateMachine.Transitions.First(t => t.Input == "ToB"); // Transition from StateA to StateB
             var oldTargetState = transitionToModify.EndState;
-            var newTargetState = sm.States.First(s => s.Name == "StateA");
+            var newTargetState = _stateMachine.States.First(s => s.Name == "StateA");
             transitionToModify.EndState = newTargetState;
-            await Task.Delay(50);
+            await Task.Delay(200);
 
-            IStateMachine parserRootSm = (IStateMachine)parser.Context.Root;
+            var parserRootSm = (IStateMachine)_parser.Context.Root;
             var modifiedTransitionInParser = parserRootSm.Transitions.First(t => t.Input == "ToB");
             Assert.That(modifiedTransitionInParser.EndState, Is.SameAs(newTargetState));
-            Assert.That(modifiedTransitionInParser.EndState, Is.Not.SameAs(oldTargetState));
 
 
-            Assert.That(parserRootSm, Is.EqualTo(sm));
-            Assert.That(parser.Context.Errors, Is.Empty);
+            Assert.That(parserRootSm, Is.EqualTo(_stateMachine));
+            _parser.Context.TryGetReferences(newTargetState, out var refsA);
+            Assert.That(refsA.Count, Is.EqualTo(3));
+            _parser.Context.TryGetReferences(oldTargetState, out var refsB);
+            Assert.That(refsB.Count, Is.EqualTo(2));
+            
+            Assert.That(_parser.Context.Errors, Is.Empty);
         }
 
         [Test]
         public async Task TestCreateStateAsync()
         {
-            var (sm, parser) = SetupStateMachineParser();
-
-            var originalStateCount = sm.States.Count;
+            var originalStateCount = _stateMachine.States.Count;
             var newName = "NewStateD";
 
             var newState = new State { Name = newName, IsStartState = false, IsEndState = false };
-            sm.States.Add(newState);
-            await Task.Delay(50);
+            _stateMachine.States.Add(newState);
+            await Task.Delay(100);
 
-            IStateMachine parserRootSm = (IStateMachine)parser.Context.Root;
+            var parserRootSm = (IStateMachine)_parser.Context.Root;
             Assert.That(parserRootSm.States.Count, Is.EqualTo(originalStateCount + 1));
             Assert.That(parserRootSm.States.Any(s => s.Name == newName), Is.True);
             Assert.That(parserRootSm.States.First(s => s.Name == newName), Is.SameAs(newState));
 
-            parser.Context.TryGetDefinition(newState, out var def);
+            _parser.Context.TryGetDefinition(newState, out var def);
             Assert.That(def, Is.Not.Null);
             Assert.That(def.ContextElement, Is.SameAs(newState));
 
 
-            Assert.That(parserRootSm, Is.EqualTo(sm));
-            Assert.That(parser.Context.Errors, Is.Empty);
+            Assert.That(parserRootSm, Is.EqualTo(_stateMachine));
+            Assert.That(_parser.Context.Errors, Is.Empty);
         }
 
         [Test]
         public async Task TestCreateTransitionAsync()
         {
-            var (sm, parser) = SetupStateMachineParser();
-
-            var originalTransitionCount = sm.Transitions.Count;
-            var sourceState = sm.States.First(s => s.Name == "StateA");
-            var targetState = sm.States.First(s => s.Name == "StateC");
+            var originalTransitionCount = _stateMachine.Transitions.Count;
+            var sourceState = _stateMachine.States.First(s => s.Name == "StateA");
+            var targetState = _stateMachine.States.First(s => s.Name == "StateC");
             var newTransitionInput = "AToC";
 
-            var newTransition = new Transition { Input = newTransitionInput, StartState = sourceState, EndState = targetState };
-            sm.Transitions.Add(newTransition);
-            await Task.Delay(50);
+            var newTransition = new Transition
+                { Input = newTransitionInput, StartState = sourceState, EndState = targetState };
+            _stateMachine.Transitions.Add(newTransition);
+            await Task.Delay(200);
 
-            IStateMachine parserRootSm = (IStateMachine)parser.Context.Root;
+            var parserRootSm = (IStateMachine)_parser.Context.Root;
             Assert.That(parserRootSm.Transitions.Count, Is.EqualTo(originalTransitionCount + 1));
             Assert.That(parserRootSm.Transitions.Any(t => t.Input == newTransitionInput), Is.True);
             Assert.That(parserRootSm.Transitions.First(t => t.Input == newTransitionInput), Is.SameAs(newTransition));
 
-            parser.Context.TryGetDefinition(newTransition, out var def);
+            _parser.Context.TryGetDefinition(newTransition, out var def);
             Assert.That(def, Is.Not.Null);
             Assert.That(def.ContextElement, Is.SameAs(newTransition));
 
 
-            Assert.That(parserRootSm, Is.EqualTo(sm));
-            Assert.That(parser.Context.Errors, Is.Empty);
+            Assert.That(parserRootSm, Is.EqualTo(_stateMachine));
+            Assert.That(_parser.Context.Errors, Is.Empty);
+        }
+
+        [Test]
+        public async Task TestReplaceStateAsync()
+        {
+            //Index of StateD
+            var index = 3;
+            var newState = new State { Name = "ReplacedState", IsStartState = false, IsEndState = false };
+            var originalStateCount = _stateMachine.States.Count;
+
+            var statesList = (IList<IState>)_stateMachine.States;
+            statesList[index] = newState;
+
+            await Task.Delay(500);
+
+            var parserRootSm = (IStateMachine)_parser.Context.Root;
+            Assert.That(_parser.Context.TryGetDefinition(newState, out _), Is.True);
+            Assert.That(_parser.Context.Input.Any(s => s.Contains("ReplacedState")), Is.True);
+            Assert.That(_parser.Context.Input.Any(s => s.Contains("transition")), Is.True);
+            Assert.That(parserRootSm.States.Count, Is.EqualTo(originalStateCount));
+            Assert.That(parserRootSm.States.Any(s => s.Name == newState.Name), Is.True);
+            Assert.That(_parser.Context.Errors, Is.Empty);
+        }
+
+        [Test]
+        public async Task TestResetCollectionAsync()
+        {
+            _stateMachine.Transitions.Clear();
+
+            await Task.Delay(200);
+
+
+            var parserRootSm = (IStateMachine)_parser.Context.Root;
+            Assert.That(_parser.Context.Input.Any(s => s.Contains("ToB")), Is.False);
+            Assert.That(_parser.Context.Input.Any(s => s.Contains("ToC")), Is.False);
+            Assert.That(_parser.Context.Input.Any(s => s.Contains("StateD")), Is.True);
+            Assert.That(_parser.Context.Input.Any(s => s.Contains("transition")), Is.False);
+            Assert.That(parserRootSm.Transitions, Is.Empty);
+            Assert.That(_parser.Context.Errors, Is.Empty);
         }
     }
 }
